Add id to (update/destroy) comment route, move StoryController to Api folder
