+++ app/Http/Controllers/Api/StoryController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
 use App\Story;
 use Illuminate\Http\Request;
